prettierrc and change file name
