+++ README.md
@@ -1,4 +1,5 @@
 # fqm-bulk-utils
+
 Utilities for handling bulk capabilities in the context of HL7® FHIR®-based measure calculation
 
 ## License
