+++ eslint.config.mjs
@@ -4,11 +4,6 @@ import eslint from '@eslint/js';
 import tseslint from 'typescript-eslint';
 import prettierConfig from 'eslint-config-prettier';
 
-export default tseslint.config(
-  eslint.configs.recommended,
-  ...tseslint.configs.recommended,
-  prettierConfig,
-  {
-   ignores: ['build']
-  }
-);
+export default tseslint.config(eslint.configs.recommended, ...tseslint.configs.recommended, prettierConfig, {
+  ignores: ['build']
+});
