+++ tsconfig.json
@@ -1,23 +1,23 @@
 {
-    "exclude": ["build", "node_modules" ],
-    "include": ["src"],
-    "compilerOptions": {
-      "resolveJsonModule": true,
-      "target": "ESNext",
-      "lib": ["ESNext"],
-      "module": "CommonJS",
-      "allowJs": true,
-      "outDir": "./build",
-      "rootDir": "./src",
-      "strict": true,
-      "esModuleInterop": true,
-      "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true,
-      "declaration": true,
-      "sourceMap": true,
-      "moduleResolution": "node"
-    },
-    "ts-node": {
-      "files": true
-    }
+  "exclude": ["build", "node_modules"],
+  "include": ["src"],
+  "compilerOptions": {
+    "resolveJsonModule": true,
+    "target": "ESNext",
+    "lib": ["ESNext"],
+    "module": "CommonJS",
+    "allowJs": true,
+    "outDir": "./build",
+    "rootDir": "./src",
+    "strict": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "declaration": true,
+    "sourceMap": true,
+    "moduleResolution": "node"
+  },
+  "ts-node": {
+    "files": true
   }
+}
